Add default data and readme
Some random data from open neuro is being used as default data for
the app to load in on init. The data I'm using is openly available
from open neuro, but the license section on their page has not
been filed in, so if there's a problem I can always find some other
data.