Add crosshairs
All canvases are must now be written oven with a fresh image with
each update so that the old crosshairs are drawn over. This can
still be done efficiently by storting the pixel of the last re write
and if it does not need updating the slice from that axis does not
need to be re fetched.